XMLHttpRequest/abort-event-order.htm

<!DOCTYPE html>
<html>
<head>
    <link rel="help" href="http://dvcs.w3.org/hg/xhr/raw-file/tip/Overview.html#the-abort()-method" data-tested-assertations="following-sibling::ol/li[4]/ol/li[5] following-sibling::ol/li[4]/ol/li[6] following-sibling::ol/li[4]/ol/li[7]/ol/li[3] following-sibling::ol/li[4]/ol/li[7]/ol/li[4]" />
    <script src="/resources/testharness.js"></script>
    <title>XMLHttpRequest: The abort() method: abort and loadend events</title>
</head>

<body>
    <div id="log"></div>

    <script type="text/javascript">
        var test = async_test();

        test.step(function()
        {
            var xhr = new XMLHttpRequest();
            var expect = [4, "abort", "loadend", "upload.abort", "upload.loadend"];
            var actual = [];

            xhr.onreadystatechange = function()
            {
                test.step(function()
                {
                    var readyState = xhr.readyState;
                    
                    if (readyState == 1)
                    {
                        xhr.abort();
                    }
                    else if (readyState == 4)
                    {
                        actual.push(readyState);
                    }
                });
            };
            
            xhr.onloadend          = function(e){ actual.push(e.type); };
            xhr.onabort            = function(e){ actual.push(e.type); };

            xhr.upload.onloadend   = function(e){ actual.push("upload." + e.type); VerifyResult() };
            xhr.upload.onabort     = function(e){ actual.push("upload." + e.type); };

            function VerifyResult()
            {
                test.step(function()
                {
                    assert_array_equals(actual, expect);
                    test.done();
                });
            };

            xhr.open("POST", "./resources/content.py", true);
            xhr.send("Test Message");
        });
    </script>
</body>
</html>
